keep changes

+++ backend/src/fixedcostsRepo.go
@@ -7,22 +7,23 @@ import (
 )
 
 type FixedCost struct {
-	ID       int `gorm:primary_key`
-	Name     string
-	Amount   int
-	From     *YearMonth
-	To       *YearMonth
-	DueMonth months `gorm:"type:string"`
+	ID        int `gorm:primary_key`
+	Name      string
+	Amount    int
+	From      *YearMonth
+	To        *YearMonth
+	DueMonth  months `gorm:"type:string"`
+	FinanceID int
 }
 
 type months []int
 
 var ALL_MONTHS = []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12}
 
-func LoadFixedCosts() *[]FixedCost {
-	//return costs
+func LoadFixedCosts(financeId int) *[]FixedCost {
+
 	var costs []FixedCost
-	DB.Find(&costs)
+	DB.Where("finance_id = ?", financeId).Find(&costs)
 
 	return &costs
 }
@@ -32,12 +33,14 @@ func SaveFixedObject(cost *FixedCost) {
 	if cost.ID == 0 {
 		DB.Create(cost)
 	} else {
+		checkPermissionFixedCost(cost.ID, cost.FinanceID)
 		DB.Save(cost)
 	}
 
 }
 
-func DeleteFixedCost(id int) {
+func DeleteFixedCost(id int, financeId int) {
+	checkPermissionFixedCost(id, financeId)
 	DB.Delete(&FixedCost{}, id)
 }
 
@@ -74,4 +77,14 @@ func (this months) Value() (driver.Value, error) {
 
 	return result, nil
 
+}
+
+func checkPermissionFixedCost(id int, financeId int) {
+	var cost FixedCost
+
+	DB.Where("id = ?", id).First(&cost)
+
+	if cost.FinanceID != financeId {
+		panic("access to costs not in your finance not allowed")
+	}
 }
